+++ Sources/AnyDoor/Views/Onboarding/OnboardingView.swift
@@ -128,6 +128,8 @@ struct OnboardingView: View {
         case .hyperKey:         OnboardingHyperKeyStep()
         case .capture:          OnboardingCaptureStep()
         case .clipboardPalette: OnboardingClipboardPaletteStep()
+        case .translation:      OnboardingTranslationStep()
+        case .imageConversion:  OnboardingImageConversionStep()
         case .customize:        OnboardingCustomizeStep(onClose: onClose)
         }
     }